requirements_sec.txt
cmsplugin-filer==1.0.1
dj-database-url==0.4.1
Django==1.8.12
django-ajax-selects==1.4.3
django-appconf==1.0.2
django-classy-tags==0.7.2
django-cms==3.2.3
django-filer==1.1.1
django-followit==0.2.1
django-formtools==1.0
django-ipware==1.1.5
django-mptt==0.8.3
django-polymorphic==0.8.1
django-registration==2.0.4
django-reversion==1.8.7
django-sekizai==0.9.0
Django-Select2==4.3.2
django-treebeard==4.0
django-vote==1.1.2
djangocms-admin-style==1.1.0
djangocms-column==1.6.0
djangocms-forms==0.1.17
djangocms-googlemap==0.4.0
djangocms-inherit==0.1.1
djangocms-installer==0.8.7
djangocms-link==1.7.2
djangocms-style==1.7.0
djangocms-text-ckeditor==2.9.1
easy-thumbnails==2.3
hashids==1.1.0
html5lib==0.9999999
jsonfield==1.0.3
Pillow==3.2.0
pytz==2016.4
requests==2.9.1
simplejson==3.8.2
six==1.10.0
tablib==0.11.2
tzlocal==1.2.2
Unidecode==0.4.19
